enterprise code seperation from advance model code and pep8 error fix.
